Sorting and translating to 3rd parties web apps
When developers are sampling local AI benchmarks, no translation is done so that timings are not corrupted by it.
When users ask for sorted information about the fastest local AI options for a certain browser and OS, then translation to browser language is done.
Using Chrome built-in AI the translation has no or little effect on users' main local AI tasks specially with constrained local resources. 
Translation on the cloud also avoids secondary effects but it is slower and/or costly.